+++ src/lib/fetcher.ts
@@ -1,4 +1,4 @@
-import getNewAccessToken from '@/lib/token';
+import getNewAccessToken from '@/lib/getNewToken';
 
 const BASE_URL = process.env.NEXT_PUBLIC_API_URL;
