+++ src/lib/uploadImage.ts
@@ -1,5 +1,5 @@
 import Cookies from 'js-cookie';
-import getNewAccessToken from '@/lib/token';
+import getNewAccessToken from '@/lib/getNewToken';
 
 const BASE_URL = process.env.NEXT_PUBLIC_API_URL;
